Remove some complexity from KcqlSettings where the functionality is only used for one connector.
