+++ kafka-connect-cassandra/src/main/scala/io/lenses/streamreactor/connect/cassandra/config/CassandraSettings.scala
@@ -211,6 +211,7 @@ object CassandraSettings extends StrictLogging {
       defaultValueStrategy,
     )
   }
+
 }
 
 object DefaultValueServeStrategy extends Enumeration {
