+++ kafka-connect-redis/src/test/scala/io/lenses/streamreactor/connect/redis/sink/config/RedisSinkSettingsTest.scala
@@ -15,8 +15,8 @@
  */
 package io.lenses.streamreactor.connect.redis.sink.config
 
-import io.lenses.streamreactor.common.rowkeys.StringGenericRowKeyBuilder
-import io.lenses.streamreactor.common.rowkeys.StringStructFieldsStringKeyBuilder
+import io.lenses.streamreactor.connect.redis.sink.rowkeys.StringGenericRowKeyBuilder
+import io.lenses.streamreactor.connect.redis.sink.rowkeys.StringStructFieldsStringKeyBuilder
 import io.lenses.streamreactor.connect.redis.sink.support.RedisMockSupport
 import org.apache.kafka.common.config.ConfigException
 import org.scalatest.matchers.should.Matchers
